SOLR-14216: Exclude HealthCheck from authentication (WIP)
